+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -1167,7 +1167,8 @@ org.apache.ignite.internal.processors.cache.persistence.tree.CorruptedTreeExcept
 org.apache.ignite.internal.processors.cache.persistence.tree.io.DataPageIO$EntryPart
 org.apache.ignite.internal.processors.cache.persistence.tree.reuse.LongListReuseBag
 org.apache.ignite.internal.processors.cache.persistence.wal.AbstractWalRecordsIterator
-org.apache.ignite.internal.processors.cache.persistence.wal.AbstractWalRecordsIterator$StartSeekingFilter
+org.apache.ignite.internal.processors.cache.persistence.wal.FileWalRecordsIterator
+org.apache.ignite.internal.processors.cache.persistence.wal.FileWalRecordsIterator$StartSeekingFilter
 org.apache.ignite.internal.processors.cache.persistence.wal.FileWriteAheadLogManager$1
 org.apache.ignite.internal.processors.cache.persistence.wal.FileWriteAheadLogManager$FileCompressorWorker$1
 org.apache.ignite.internal.processors.cache.persistence.wal.FileWriteAheadLogManager$RecordsIterator
